move pivot into shooter folder
