+++ src/main/java/frc/robot/subsystems/shooter/pivot/PivotIOSparkMax.java
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot.subsystems.pivot;
+package frc.robot.subsystems.shooter.pivot;
 
 import static frc.robot.subsystems.intake.IntakeConstants.angleMotorId;
